Implement some basic next page decision logic
There need to be some set rules about how branching works. This commit
gets us to a point where these can be experimented with by implementing
Wizardry::Routing::NextPage, a class that can be initialized with the
name of a next page and a condition.

If the NextPage is unconditional, that page will be the next page. If
there's a condition attached, it will only be the next page if the
condition evaluates to true (in the context of the wizard object).

Refs #3