Nest branched pages in their parent
Having to distinguish between trunk and branch pages when they're on the
same level of the hierarchy is annoying and makes it difficult to work
out what order things should happen in.

The previous logic meant that branch steps (b) belonged to the previous
trunk page.

t---t---[b]---[b]---t---t---

Now, the branch pages are nested inside the trunk page in the wizard
definition (and placed inline when the wizard's initialized, invisibly
to the developer).

t---[t---]--t-------t-------
    |    |
    b----b

This makes defining the order and ownership much easier, and also
alleviates the need to define `branch: true` on initialize. Any 'sub
page' (page defined inside a `pages:` arg) will automatically be marked
as a branch page.

We can probably extend this approach later for loop/collection pages.