Add sense checks to warn on multiple special pages
There should probably only be one check your answers and completion page
per wizard. My current thinking is that paths should converge for these
pages and, if there are multiple 'exits', the page contents can be
customised in the partial.

This might change later and there's definitely a possibilty having
multiples is the right answer, so we'll just warn for now rather than
erroring