Raise error when interacting with completed wizard
The 'completed' state of a wizard needs to be explicitly controlled via
a flag rather than by just checking that all the necessary fields have
been filled in because we need to differentiate between being at the
'check your answers' stage and having confirmed our answers are correct.

The intention is to use a boolean field that defaults to :complete but
can be overridden when initializing the wizard. The submission of the
check your answers page will toggle it to true.

Refs #4